made the compared objects same in type

+++ test/edu/jsu/mcis/cs310/tas_fa22/PunchCreateTest.java
@@ -54,7 +54,7 @@ public class PunchCreateTest {
 
         rts = p2.getOriginaltimestamp();
 
-        assertEquals(terminalid, p2.getTerminalid());
+        assertEquals(terminalid, p2.getTerminalid().intValue());
         assertEquals(punchtype, p2.getPunchtype());
         assertEquals(ots.format(dtf), rts.format(dtf));
